removing lwan-syslog.[ch]
