+++ src/lib/CMakeLists.txt
@@ -25,7 +25,6 @@ set(SOURCES
 	lwan-response.c
 	lwan-socket.c
 	lwan-status.c
-	lwan-syslog.c
 	lwan-straitjacket.c
 	lwan-strbuf.c
 	lwan-tables.c
